--- a/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets
+++ b/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets
@@ -2182,6 +2182,7 @@ because xbuild doesn't support framework reference assemblies.
       ApkOutputPath="$(_ApkOutputPath)"
       FilesToAddToArchive="@(FilesToAddToArchive)"
       UncompressedFileExtensions="$(AndroidStoreUncompressedFileExtensions)"
+      UseLibZipSharp="$(_AndroidUseLibZipSharp)"
       ZipFlushFilesLimit="$(_ZipFlushFilesLimit)"
       ZipFlushSizeLimit="$(_ZipFlushSizeLimit)" />
 
@@ -2313,6 +2314,7 @@ because xbuild doesn't support framework reference assemblies.
       ApkOutputPath="$(_ApkOutputPath)"
       FilesToAddToArchive="@(FilesToAddToArchive)"
       UncompressedFileExtensions="$(AndroidStoreUncompressedFileExtensions)"
+      UseLibZipSharp="$(_AndroidUseLibZipSharp)"
       ZipFlushFilesLimit="$(_ZipFlushFilesLimit)"
       ZipFlushSizeLimit="$(_ZipFlushSizeLimit)" />