… src/Xamarin.Android.Build.Tasks/Utilities/ZipArchiveFileListBuilder.cs