@dellis1972 2024-12-17
One of the reasons we didn't use System.IO.Compression was because it IGNORED per file compression settings it also failed to actually STORE things when NoCompression was defined (it would always compress files), this would break out native library support in all sorts of weird ways. 

If the plan is to use this when we are going to need to make sure the test coverage is making sure we can add files which have no compression and that they are actually stored as such.

//cc @grendello