@jonathanpeppers 2024-12-17
@dellis1972 I think he figured out how to set the compression method now:
```
netstandard2.0 does not expose API for examining a CRC or CompressionMethod of an existing entry in a Zip file, though both net472 and net9.0 have private fields. As such, we use reflection to access these private fields. If the runtime we are using does not have these fields, we will fall back to using LibZipSharp as we do now.
```