@jpobst 2024-12-17
> One of the reasons we didn't use System.IO.Compression was because it IGNORED per file compression settings it also failed to actually STORE things when NoCompression was defined (it would always compress files), this would break out native library support in all sorts of weird ways.

Thanks for the context.  I'll do some testing around `NoCompression` to verify what .NET actually does.