@jonathanpeppers 2024-12-17
@jpobst it might be enough to verify in some of our build tests that the compression method is correct.

We have a couple like this:

https://github.com/dotnet/android/blob/de043164416222ebf703f04205279fdcac56a9af/tests/MSBuildDeviceIntegration/Tests/BundleToolTests.cs#L299-L307