@jpobst 2024-12-17
I believe I see the issue that @dellis1972 is referring to.

Building with `NoCompression` on .NET 9 looks correct:
- Size matches Packed Size
- Method is `Store`

![image](https://github.com/user-attachments/assets/0c7e00ca-20b8-4474-b5fd-13beb2358795)

However, using `NoCompression` on .NET Framework 4.7.2:
- The contents are "uncompressed" but the sizes are slightly off
- Method is `Deflate`

![image](https://github.com/user-attachments/assets/ba22765c-4e28-4d23-b1e7-a7a88871acd7)

It's like it's running it through the compression algorithm with a compression level of 0 instead of placing it unmodified inside the archive.

I think we have 2 options here:
- Use `LibZipSharp` when running inside VS Win (.NET Framework).
- Move the compression out of process so we can always run on .NET "Core".