@jonathanpeppers 2024-12-17
Could we include the source from the “good” System.IO.Compression in src-ThirdParty? Then we could also avoid System.Reflection by making some new public member.