@jpobst 2024-12-17
> Could we include the source from the “good” System.IO.Compression in src-ThirdParty? Then we could also avoid System.Reflection by making some new public member.

I think the biggest issue is we need it to run on `netstandard2.0` in order to be usable by VS MSBuild.  It is an unknown, but this could be non-trivial to port from `net9.0`.  Also, the modern .NET version uses `zlib`, which .NET redistributes on Windows because it isn't provided by the OS, so we would be on the hook for that too.

I think we would likely be venturing into undesirable territory trying to ship our own.

---

If we moved it out-of-process, we could eliminate the CRC reflection because a new [property exists](https://learn.microsoft.com/en-us/dotnet/api/system.io.compression.ziparchiveentry.crc32?view=net-9.0).  There still isn't a property for the compression method, but we could possibly apply some pressure on [this issue](https://github.com/dotnet/runtime/issues/95909) for .NET 10.

(We use the CRC and CompressionMethod to check if we need to update existing entries in an archive during incremental builds.)