@jonathanpeppers 2024-12-18
> modern .NET version uses zlib,

Yeah, so I wouldn't try to bring their code in.

Maybe out-of-process is a good option? They are doing this for components inside Visual Studio as well. A new console app could do the work of the new target, and could always run under .NET 10+ and not .NET framework at all.