@jpobst 2025-01-07
It appears that our usage of `$(AndroidStoreUncompressedFileExtensions)` is pretty limited these days.  It is not used (by default) in any of the common cases I am testing:

- Debug - FastDev - empty
- Debug - EmbedAssembliesInApk - empty
- Release - Specifies `.blob`, but no files match that specification (they are now stored as `.blob.so`)

It looks like it is [only used with](https://github.com/dotnet/android/blob/e0bf801aea28d18ce735dc0fbc1c5546dac173fb/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Common.targets#L1601-L1605):

- `extractNativeLibs` = `false`
- `useEmbeddedDex` = `true`
- User explicitly specifies file extensions

Given this, I think the best course of action is simply to determine if any file is actually going to be stored uncompressed and fall back to LibZipSharp if needed and running on .NET Framework.

This gives us the performance win in the majority of cases without complex workarounds.