@jpobst 2025-01-08
> One thing to note about extractNativeLibs = false, is that app bundles likely will use this.

When `$(AndroidPackageFormat)` = `aab`, we already ignore `$(AndroidStoreUncompressedFileExtensions)` and compress everything, so it will use `System.IO.Compression`.