@grendello 2025-01-13
> @grendello can you remember any other issues we had with the zip files System.IO.Compression produced? I seem to recall is wasn't just the `Store` issue.

Unix permissions were all botched, we'd sometimes end up with files that had permission of `000` (completely unreadable)