@jpobst 2025-01-13
> Unix permissions were all botched, we'd sometimes end up with files that had permission of 000 (completely unreadable)

Looks like this is addressed here: https://github.com/dotnet/runtime/pull/70735.  It sounds like it was only an issue when creating zip files on Unix, which would always be .NET 10+ (containing the fix) for us.

If the fix is needed on Windows, we could possibly use [ZipArchiveEntry.ExternalAttributes](https://learn.microsoft.com/en-us/dotnet/api/system.io.compression.ziparchiveentry.externalattributes?view=net-9.0#system-io-compression-ziparchiveentry-externalattributes) to set it manually: https://github.com/dotnet/runtime/issues/17912#issuecomment-641594638.